Implement Delete Functionality, User Listing, and Open User Profile
This update introduces several important features, including the ability to delete users from the system, improving overall data management. In addition to the user listing, which displays all registered users and facilitates navigation through the user base, a feature has also been added that redirects to the user's Instagram link when clicking to open a user profile. This allows for easier access to social media.